+++ Native/Account/app/src/main/java/com/example/account/utils/Helper.kt
@@ -62,7 +62,7 @@ fun getItemTotal(item: InvoiceItem): String {
     )
 }
 
-fun getItemTotal(price: Double, quantity: Int): String {
+fun getItemTotal(price: Float, quantity: Int): String {
     val total = price * quantity
     return total.toBigDecimal().setScale(2, RoundingMode.HALF_DOWN).toDouble().toString()
 }
